requirements.txt
Flask==2.3.2
opencv-python==4.9.0
torch==2.0.1
torchvision==0.15.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
scipy==1.10.1
numpy==1.24.3
pyyaml==6.0
tqdm==4.65.0
matplotlib==3.7.1
pandas==2.0.3